backend/requirements/base.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
cryptography==42.0.5
python-dotenv==1.0.1
gunicorn==21.2.0
django-filter==24.1
Pillow==10.2.0
pyotp==2.9.0 